Refactor the handlers and move some methods to context.
